+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 [comment]: # ( Copyright Contributors to the Open Cluster Management project )
 # Release Content
 ## Additions
-- version 1.0.0 
+- Display message "clusterclaim xx is running" only when it wasn't running before.
 ## Breaking changes
 
 ## Bug fixes